user-managment-service/requirements.txt
fastapi
uvicorn
aiohttp
boto3
python-dotenv
pandas
sqlalchemy
tenacity
aiomysql
mysqlclient
asyncpg